Curating the source code
Changed the code, so that it's fully Zeppeling SafeMath-compliant

Also, tweaked the ratios so that it actually settle for the right block-time. After further testings, I figured out that the difficulty algorithm doesn't take into account the time to submit the transaction on the network.

Now, it does take that into account.